+++ src/main/java/com/trioshop/controller/user/UserController.java
@@ -21,59 +21,6 @@ public class UserController {
     @Autowired
     private UserInfoService userInfoService;
 
-    @GetMapping("/login")
-    public String loginPage_G(@RequestParam(value = "error", required = false)String error, Model model) {
-        if(Objects.nonNull(error)){
-            model.addAttribute("error","정보를 찾을 수 없습니다.");
-        }
-        return "/user/userInfo/login";
-    }
-
-    //@ModelAttribute 객체를 받아오는거 @RequestParam 변수명을 가져오는거
-    @PostMapping("/login")
-    public String loginPage(@ModelAttribute UserIdPasswd userIdPasswd) {
-        UserInfoBySession user = userInfoService.isValidUser(userIdPasswd);
-        String redirectURI="/";
-        session.setAttribute(SessionConst.LOGIN_MEMBER, user);
-        //스프링이 자동으로 관리하는 세션 객체에 속성이 설정됩니다. 이렇게 하면 사용자가 로그인한 정보를 세션에 저장할 수 있습니다.
-        if (user.getGradeCode() == 4) {
-           redirectURI = "/trioAdmin";
-        }
-        return "redirect:"+redirectURI;
-    }
-
-
-    @GetMapping("/logout")
-    public String logoutPage(HttpSession session) {
-        session.invalidate();
-        return "redirect:/";
-    }
-
-    //@ModelAttribute 객체로 반환 UserJoin객체를 userJoin 라는 이름으로 가져온것임
-    @GetMapping("/join")
-    public String joinPage_G(@ModelAttribute("userJoin") UserJoin userJoin) {
-        return "/user/userInfo/join";
-    }
-
-    @PostMapping("/join")
-    public ModelAndView joinPage(@ModelAttribute("userJoin") UserJoin userJoin) {
-        ModelAndView mv = new ModelAndView();
-        try {
-            boolean isRegistered = userInfoService.saveUserInfo(userJoin);
-            if (isRegistered) {
-                mv.setViewName("redirect:/login");
-                mv.addObject("success", "회원가입에 성공했습니다.");
-            } else {
-                mv.setViewName("redirect:/join");
-                mv.addObject("error", "이미 사용중인 계정입니다.");
-            }
-        } catch (Exception e) {
-            mv.setViewName("redirect:/join");
-            mv.addObject("exception", "회원가입 중 오류가 발생했습니다.");
-        }
-        return mv;
-    }
-
     @GetMapping("/findId")
     public String findId_G() {
         return "/user/userInfo/findId";
@@ -206,43 +153,5 @@ public class UserController {
         }
     }
 
-    //@ModelAttribute 폼에서입력하면 컨트롤러로 전달~~
-    @GetMapping("/guestLogin")
-    public String guestLoginPage() {
-        return "/user/userInfo/guestLogin";
-    }
 
-    @PostMapping("/guestLogin")
-    public ModelAndView guestLogin(@ModelAttribute GuestUserJoin guestUserJoin, @ModelAttribute GuestUserJoin2 guestUserJoin2) {
-        ModelAndView mv = new ModelAndView();
-
-        // 첫 번째 로그인 시도
-        GuestUserJoin existingUser = userInfoService.LoginGuestUser(guestUserJoin);
-
-        // 기존 사용자가 있고 grade_code가 0인 경우에만 로그인 성공
-        if (existingUser != null && existingUser.getGradeCode() == 0) {
-            mv.setViewName("redirect:/");
-        } else {
-            // guestUserJoin 객체에 필요한 값 설정
-            guestUserJoin.setGradeCode(0); // 예시로 gradeCode를 설정하고, 필요한 다른 값들도 설정해야 함
-
-            // 중복된 DB가 없으면 회원가입을 시도
-            boolean isSuccess = userInfoService.saveGuestUser(guestUserJoin, guestUserJoin2);
-            if (isSuccess) {
-                // 회원가입이 완료되면 자동으로 로그인
-                mv.addObject("message", "회원가입이 완료되었습니다. 로그인되었습니다.");
-            } else {
-                mv.addObject("message", "로그인에 실패했습니다. 다시 시도해주세요.");
-            }
-            mv.setViewName("redirect:/");
-        }
-        // 나중에 수정해야할 부분
-        UserInfoBySession sessionUser = new UserInfoBySession();
-        sessionUser.setUserNickname("게스트유저");
-        sessionUser.setUserCode(guestUserJoin.getUserCode());
-        sessionUser.setGradeCode(1);
-        session.setAttribute(SessionConst.LOGIN_MEMBER, sessionUser);
-        ////수정
-        return mv;
-    }
 }
